+++ src/compute-engine/rules.ts
@@ -10,58 +10,330 @@ import {
   SemiBoxedExpression,
   PatternReplaceFunction,
 } from './public';
-import { asLatexString } from './boxed-expression/utils';
+import {
+  asLatexString,
+  isInequality,
+  isRelationalOperator,
+} from './boxed-expression/utils';
+import { isCollection } from './collection-utils';
+import { Parser } from './latex-syntax/public';
+import { isRational } from './numerics/rationals';
 
-/**
- * For each rules in the rule set that match, return the `replace` of the rule
- *
- * @param rules
- */
-export function matchRules(
-  expr: BoxedExpression,
-  rules: BoxedRuleSet,
-  sub: BoxedSubstitution
-): BoxedExpression[] {
-  const results: BoxedExpression[] = [];
-  for (const rule of rules) {
-    const r = applyRule(rule, expr, sub);
-    if (r === null) continue;
-    // Verify that the results are unique
-    if (results.some((x) => x.isSame(r))) continue;
-    results.push(r);
-  }
-
-  return results;
-}
-
-// @future Generator functions
+// @todo ['Alternatives', ...]:
+// @todo: ['Condition',...] : Conditional match
+// @todo: ['Repeated',...] : repeating match
+// @todo _x:Head or _x:RealNumbers
+// @todo Generator functions
 // export function fixPoint(rule: Rule);
 // export function chain(rules: RuleSet);
 
-// @future load rules from JSONC
-// - describe conditions with a condition expression:
-//    "x.isInteger && y.isFreeOf('x')"
-//  or"x:integer && y:freeOf(x)"
-// function parseCondition(s:string, lhs: (sub)=> boolean) =>
-//    [rest: string, fn: (sub) => boolean]
+// Conditions:
+// :boolean       - a boolean value, True or False
+// :string        - a string of characters
+// :number        - a number literal
+// :symbol
+// :expression
 
-// @future: priority for rules, sort and apply rules by priority
+// :numeric       - an expression that has a numeric value, i.e. 2√3, 1/2, 3.14
+// :integer       - an integer value, -2, -1, 0, 1, 2, 3, ...
+// :natural       - a natural number, 0, 1, 2, 3, ...
+// :\Z^-          - a negative integer, -1, -2, -3, ...
+// :\Z^+          - a positive integer, 1, 2, 3, ...
+// :real          - real numbers, including integers
+// :imaginary     - imaginary numbers, i.e. 2i, 3√-1  (not including real numbers)
+// :complex       - complex numbers, including real and imaginary
+// :rational      - rational numbers, 1/2, 3/4, 5/6, ...
+// :irrational    - irrational numbers, √2, √3, π, ...
+// :algebraic     - algebraic numbers, rational and irrational
+// :transcendental  - transcendental numbers, π, e, ...
 
-/**
- * Create a boxed rule set from a non-boxed rule set
- * @param ce
- * @param rs
- * @returns
- */
-export function boxRules(ce: IComputeEngine, rs: Iterable<Rule>): BoxedRuleSet {
-  const result: BoxedRule[] = [];
+// :positive      - positive real numbers, > 0
+// :negative      - negative real numbers, < 0
+// :nonnegative   - nonnegative real numbers, >= 0
+// :nonpositive   - nonpositive real numbers, <= 0
 
-  for (const rule of rs) result.push(normalizeRule(ce, rule));
+// :even          - even integers, 0, 2, 4, 6, ...
+// :odd           - odd integers, 1, 3, 5, 7, ...
 
-  return result.sort((a, b) => b.priority - a.priority);
+// :prime         :A000040 - prime numbers, 2, 3, 5, 7, 11, ...
+// :composite     :A002808 - composite numbers, 4, 6, 8, 9, 10, ...
+
+// :notzero       - a value that is not zero
+// :notone        - a value that is not one
+
+// :finite        - a finite value, not infinite
+// :infinite
+
+// :constant
+// :variable
+
+// :function
+
+// :operator
+// :relation      - an equation or inequality
+// :equation
+// :inequality    -
+
+// :vector        - a tensor of rank 1
+// :matrix        - a tensor of rank 2
+// :list          - a collection of values
+// :set           - a collection of unique values
+// :tuple         - a fixed length list
+// :single        - a tuple of length 1
+// :pair          - a tuple of length 2
+// :triple        - a tuple of length 3
+// :collection    - a list, set, or tuple
+// :tensor        - a nested list of values of the same type
+// :scalar        - not a tensor or list
+
+// :unit
+// :dimension
+// :angle
+// :polynomial    - an expression that is a sum of terms
+// :A000000       - a number belonging to a sequence in the OEIS
+
+export const ConditionParent = {
+  boolean: '',
+  string: '',
+  expression: '',
+
+  numeric: 'expression',
+  number: 'numeric',
+  symbol: 'expression',
+
+  complex: 'number',
+  imaginary: 'complex',
+  real: 'complex',
+  integer: 'real',
+  rational: 'real',
+  irrational: 'real',
+  algebraic: 'real',
+  transcendental: 'real',
+
+  notzero: 'number',
+  notone: 'number',
+
+  finite: 'number',
+  infinite: 'number',
+
+  positive: 'real',
+  negative: 'real',
+  nonnegative: 'real',
+  nonpositive: 'real',
+
+  even: 'integer',
+  odd: 'integer',
+
+  prime: 'integer',
+  composite: 'integer',
+
+  constant: 'expression',
+  variable: 'expression',
+
+  function: 'expression',
+
+  operator: 'expression',
+  relation: 'operator',
+  equation: 'relation',
+  inequality: 'relation',
+
+  collection: 'expression',
+  list: 'collection',
+  set: 'collection',
+
+  tuple: 'collection',
+  single: 'tuple',
+  pair: 'tuple',
+  triple: 'tuple',
+
+  tensor: 'collection',
+  vector: 'tensor',
+  matrix: 'tensor',
+  scalar: 'expression',
+
+  unit: 'expression',
+  dimension: 'expression',
+  angle: 'expression',
+  polynomial: 'expression',
+};
+
+export const CONDITIONS = {
+  boolean: (x: BoxedExpression) => x.domain?.isCompatible('Booleans'),
+  string: (x: BoxedExpression) => x.string !== null,
+  number: (x: BoxedExpression) => x.numericValue !== null,
+  symbol: (x: BoxedExpression) => x.symbol !== null,
+  expression: (x: BoxedExpression) => true,
+
+  numeric: (x: BoxedExpression) => {
+    const [c, term] = x.engine._toNumericValue(x);
+    return term.isOne;
+  },
+  integer: (x: BoxedExpression) => x.isInteger,
+  real: (x: BoxedExpression) => x.isReal,
+  complex: (x: BoxedExpression) => x.isComplex,
+  imaginary: (x: BoxedExpression) => x.isImaginary,
+  rational: (x: BoxedExpression) => x.isRational,
+  irrational: (x: BoxedExpression) => x.domain?.isRational === false,
+  algebraic: (x: BoxedExpression) => x.isAlgebraic,
+  transcendental: (x: BoxedExpression) => x.domain?.isAlgebraic === false,
+
+  positive: (x: BoxedExpression) => x.isPositive,
+  negative: (x: BoxedExpression) => x.isNegative,
+  nonnegative: (x: BoxedExpression) => x.isNonNegative,
+  nonpositive: (x: BoxedExpression) => x.isNonPositive,
+
+  even: (x: BoxedExpression) => x.isEven,
+  odd: (x: BoxedExpression) => x.isOdd,
+
+  prime: (x: BoxedExpression) => x.isPrime,
+  composite: (x: BoxedExpression) => x.isComposite,
+
+  notzero: (x: BoxedExpression) => x.isNotZero,
+  notone: (x: BoxedExpression) => !x.isOne,
+
+  finite: (x: BoxedExpression) => x.isFinite,
+  infinite: (x: BoxedExpression) => !x.isFinite,
+
+  constant: (x: BoxedExpression) => x.symbol !== null && x.isConstant,
+  variable: (x: BoxedExpression) =>
+    x.symbol !== null && !x.domain?.isFunction && !x.isConstant,
+  function: (x: BoxedExpression) => x.symbol !== null && x.domain?.isFunction,
+
+  relation: (x: BoxedExpression) => isRelationalOperator(x.head),
+  equation: (x: BoxedExpression) => x.head === 'Equal',
+  inequality: (x: BoxedExpression) => isInequality(x),
+
+  collection: (x: BoxedExpression) => isCollection(x),
+  list: (x: BoxedExpression) => x.head === 'List',
+  set: (x: BoxedExpression) => x.head === 'Set',
+  tuple: (x: BoxedExpression) =>
+    x.head === 'Tuple' ||
+    x.head === 'Single' ||
+    x.head === 'Pair' ||
+    x.head === 'Triple',
+  single: (x: BoxedExpression) => x.head === 'Single',
+  pair: (x: BoxedExpression) => x.head === 'Pair',
+  triple: (x: BoxedExpression) => x.head === 'Triple',
+  tensor: (x: BoxedExpression) => x.rank > 0,
+  vector: (x: BoxedExpression) => x.rank === 1,
+  matrix: (x: BoxedExpression) => x.rank === 2,
+  scalar: (x: BoxedExpression) => x.rank === 0,
+
+  unit: (x: BoxedExpression) => x.head === 'Unit',
+  dimension: (x: BoxedExpression) => x.head === 'Dimension',
+  angle: (x: BoxedExpression) => x.head === 'Angle',
+  polynomial: (x: BoxedExpression) => x.unknowns.length === 1,
+};
+
+function checkConditions(x: BoxedExpression, conditions: string[]): boolean {
+  // Check for !== true, because result could also be undefined
+  for (const cond of conditions) if (CONDITIONS[cond](x) !== true) return false;
+
+  return true;
 }
 
-function normalizeLatexRule(
+function tokenizeLaTeX(input: string): string[] {
+  // Regular expression to match LaTeX tokens
+  const regex = /\\[a-zA-Z]+|[{}]|[\d]+|[+\-*/^_=(),.;]|[a-zA-Z]/g;
+
+  // Match the input string against the regular expression
+  const tokens = input.match(regex);
+  // If no tokens are found, return an empty array
+  if (!tokens) return [];
+  return tokens.filter((x) => x !== ' '); // Remove spaces;
+}
+
+function parseModifier(parser: Parser): string | null {
+  // Is it a tagged modifier of the form `\mathrm{modifier}`?
+  const next = parser.peek;
+  let modifier: string | null = null;
+  if (next === '\\mathrm') {
+    parser.nextToken();
+    modifier = parser.parseStringGroup();
+  } else if (/^[a-z]$/.test(next)) {
+    // We also accept unwrapped modifiers, i.e. `:modifier`
+    modifier = parser.nextToken();
+    while (/^[a-z]$/.test(parser.peek)) modifier += parser.nextToken();
+  } else {
+    // We accept some shortcuts for common conditions
+    const shortcuts = {
+      '>0': 'positive',
+      '\\gt0': 'positive',
+      '<0': 'negative',
+      '\\lt0': 'negative',
+      '>=0': 'nonnegative',
+      '\\geq0': 'nonnegative',
+      '<=0': 'nonpositive',
+      '\\leq0': 'nonpositive',
+      '!=0': 'notzero',
+      '\\neq0': 'notzero',
+      '\\neq1': 'notone',
+      '!=1': 'notone',
+      '\\in\\Z': 'integer',
+      '\\in\\mathbb{Z}': 'integer',
+      '\\in\\N': 'natural',
+      '\\in\\mathbb{N}': 'natural',
+      '\\in\\R': 'real',
+      '\\in\\mathbb{R}': 'real',
+      '\\in\\C': 'complex',
+      '\\in\\mathbb{C}': 'complex',
+      '\\in\\Q': 'rational',
+      '\\in\\mathbb{Q}': 'rational',
+      '\\in\\Z^+': 'integer,positive',
+      '\\in\\Z^-': 'intger,negative',
+      '\\in\\Z^*': 'nonzero',
+      '\\in\\R^+': 'positive',
+      '\\in\\R^-': 'negative',
+      '\\in\\R^*': 'real,nonzero',
+      '\\in\\N^*': 'integer,positive',
+      '\\in\\N_0': 'integer,nonnegative',
+      '\\in\\R\\backslash\\Q': 'irrational',
+    };
+    for (const shortcut in shortcuts) {
+      const tokens = tokenizeLaTeX(shortcut);
+      const start = parser.index;
+      for (const token of tokens) {
+        if (!parser.match(token)) {
+          parser.index = start;
+          break;
+        }
+        parser.skipSpace();
+      }
+      modifier = shortcuts[shortcut];
+      break;
+    }
+  }
+
+  if (!modifier) return null;
+  if (!Object.keys(CONDITIONS).includes(modifier))
+    throw new Error(`Unexpected condition ${modifier}`);
+  return modifier;
+}
+
+function parserModifiers(parser: Parser): string {
+  let modifiers = '';
+  do {
+    const modifier = parseModifier(parser);
+    if (!modifier) break;
+    modifiers += modifier;
+  } while (parser.match(','));
+  return modifiers;
+}
+
+// Look for a modifier expression of the form
+// `:condition1,condition2,...`
+// or `_{condition1,condition2,...}`
+function parseModifierExpression(parser: Parser): string | null {
+  let conditions = '';
+  if (parser.match(':')) conditions = parserModifiers(parser);
+  else if (parser.matchAll(['_', '<{>'])) {
+    conditions = parserModifiers(parser);
+    if (!parser.match('<}>')) return null;
+  }
+  return conditions;
+}
+
+function parseLatexRule(
   ce: IComputeEngine,
   rule?: string | SemiBoxedExpression | PatternReplaceFunction
 ): BoxedExpression | undefined {
@@ -81,19 +353,160 @@ function normalizeLatexRule(
   return ce.box(rule, { canonical: false });
 }
 
-function normalizeStringRule(ce: IComputeEngine, rule: string): BoxedRule {
-  const [lhs, rhs] = rule.split(/->|\\to/).map((x) => x.trim());
-  return normalizeRule(ce, {
-    match: normalizeLatexRule(ce, lhs),
-    replace: normalizeLatexRule(ce, rhs)!,
+/** A rule can be expressed as a string of the form
+ * "<match> -> <replace>; <condition>"
+ * where `<match>`, `<replace>` and `<condition>` are LaTeX expressions.
+ */
+function parseRule(ce: IComputeEngine, rule: string): BoxedRule {
+  const makeWildcardEntry = (x: string) => {
+    return {
+      kind: 'symbol',
+      latexTrigger: x,
+      // domain: { kind: 'Any' },
+      parse: (parser, until) => {
+        // console.log(parser.peek);
+        if (!wildcards[x]) wildcards[x] = `_${x}`;
+        // conditions are `:condition` or `:condition1,condition2,...`
+        // or `:\mathrm{condition}`
+        let conditions = parseModifierExpression(parser);
+        if (conditions === null) return null;
+        if (conditions) {
+          if (!wildcardConditions[x]) wildcardConditions[x] = conditions;
+          else wildcardConditions[x] += ',' + conditions;
+        }
+
+        return wildcards[x];
+      },
+    };
+  };
+
+  // Setup custom dictionary entries for ...x, ...x?
+  // mapping to wildcard sequence, wildcard optional sequence
+  const previousDictionary = ce.latexDictionary;
+
+  // A mapping from an identifier to a wildcard
+  const wildcards: Record<string, string> = {};
+  // A mapping from an identifier to a condition
+  const wildcardConditions: Record<string, string> = {};
+
+  // Add wildcard entries for all lowercase letters, execpt
+  // for e (natural number), d (differential) and i (imaginary unit)
+  ce.latexDictionary = [
+    ...previousDictionary,
+    {
+      kind: 'prefix',
+      precedence: 100,
+      latexTrigger: '...',
+      parse: (parser, until) => {
+        const id = parser.nextToken();
+        if (!'abcfghjklmnopqrstuvwxyz'.includes(id)) return null;
+        let prefix = '__';
+        // Optional wildcard sequence?
+        if (parser.match('?')) prefix = '___';
+
+        if (wildcards[id] && wildcards[id] !== `${prefix}${id}`)
+          throw new Error(`Duplicate wildcard ${id}`);
+        if (!wildcards[id]) wildcards[id] = `${prefix}${id}`;
+
+        // Check for conditions
+        let conditions = parseModifierExpression(parser);
+        if (conditions === null) return null;
+
+        if (conditions) {
+          if (!wildcardConditions[id]) wildcardConditions[id] = conditions;
+          else wildcardConditions[id] += ',' + conditions;
+        }
+
+        return `${prefix}${id}`;
+      },
+    },
+    ...'abcfghjklmnopqrstuvwxyz'.split('').map(makeWildcardEntry),
+    {
+      kind: 'infix',
+      precedence: 100,
+      latexTrigger: '->',
+      parse: (parser, lhs, until) => {
+        const rhs = parser.parseExpression(until);
+        if (!rhs) return null;
+        if (parser.match(';')) {
+          // condition is either a predicate, or a sequence of wildcards + ":" + modifiers
+          // Try the sequence of wildcards first
+          let done = false;
+          let start = parser.index;
+          do {
+            const id = parser.nextToken();
+            if (wildcards[id]) {
+              let conditions = parseModifierExpression(parser);
+              if (conditions === null || !conditions) {
+                done = true;
+                parser.index = start;
+                break;
+              }
+              if (!wildcardConditions[id]) wildcardConditions[id] = conditions;
+              else wildcardConditions[id] += ',' + conditions;
+            }
+          } while (!done);
+          if (!parser.atEnd) {
+            parser.index = start;
+            const condition = parser.parseExpression(until);
+            if (condition) return ['Rule', lhs, rhs, condition];
+          }
+        }
+        // Check if we have some accumulated conditions
+        let conditions: any[] = [];
+        for (const id in wildcardConditions) {
+          const xs = wildcardConditions[id].split(',');
+          if (xs.length === 0) continue;
+          if (xs.length === 1)
+            conditions.push(['Condition', wildcards[id], xs[0]]);
+          else conditions.push(['Condition', wildcards[id], ['List', ...xs]]);
+        }
+
+        if (conditions.length === 0) return ['Rule', lhs, rhs];
+        if (conditions.length === 1) return ['Rule', lhs, rhs, conditions[0]];
+        return ['Rule', lhs, rhs, ['List', ...conditions]];
+      },
+    },
+  ];
+  let expr = ce.parse(rule, { canonical: false });
+  ce.latexDictionary = previousDictionary;
+
+  if (expr.head !== 'Rule') throw new Error(`Invalid rule ${rule}`);
+  const [match, replace, condition] = expr.ops!;
+
+  let condFn: undefined | PatternConditionFunction = undefined;
+  if (condition) {
+    condFn = (sub: BoxedSubstitution, _ce: IComputeEngine): boolean => {
+      for (const id of Object.keys(sub)) {
+        // Map the id to a wildcard
+
+        // Find a key in the wildcards that matches the id
+        const idx =
+          Object.keys(wildcards)[
+            Object.values(wildcards).findIndex((x) => x === id)
+          ];
+        if (idx === undefined) continue;
+
+        const wcCond = wildcardConditions[idx];
+        if (!wcCond) continue;
+        const conditions = wcCond.split(',');
+        if (!checkConditions(sub[id], conditions)) return false;
+      }
+      return true;
+    };
+  }
+
+  return boxRule(ce, {
+    match,
+    replace,
     priority: 0,
-    condition: undefined,
-    id: lhs + ' -> ' + rhs,
+    condition: condFn,
+    id: match.toString() + ' -> ' + replace.toString(),
   });
 }
 
-function normalizeRule(ce: IComputeEngine, rule: Rule): BoxedRule {
-  if (typeof rule === 'string') return normalizeStringRule(ce, rule);
+function boxRule(ce: IComputeEngine, rule: Rule): BoxedRule {
+  if (typeof rule === 'string') return parseRule(ce, rule);
 
   const { match, replace, condition, priority, id } = rule;
 
@@ -109,8 +522,8 @@ function normalizeRule(ce: IComputeEngine, rule: Rule): BoxedRule {
     }
   } else condFn = condition;
 
-  const matchExpr = normalizeLatexRule(ce, match);
-  const replaceExpr = normalizeLatexRule(ce, replace);
+  const matchExpr = parseLatexRule(ce, match);
+  const replaceExpr = parseLatexRule(ce, replace);
   return {
     match: matchExpr,
     replace: replaceExpr ?? (replace as PatternReplaceFunction),
@@ -119,6 +532,17 @@ function normalizeRule(ce: IComputeEngine, rule: Rule): BoxedRule {
     exact: rule.exact ?? true,
     id: id ?? (matchExpr?.latex ?? '') + ' -> ' + replaceExpr?.latex ?? '',
   };
+}
+
+/**
+ * Create a boxed rule set from a collection of non-boxed rules
+ */
+export function boxRules(ce: IComputeEngine, rs: Iterable<Rule>): BoxedRuleSet {
+  const result: BoxedRule[] = [];
+
+  for (const rule of rs) result.push(boxRule(ce, rule));
+
+  return result.sort((a, b) => b.priority - a.priority);
 }
 
 /**
@@ -136,6 +560,19 @@ function applyRule(
   options?: ReplaceOptions
 ): BoxedExpression | null {
   const { match, replace, condition, id } = rule;
+
+  if (expr.numericValue && isRational(expr.numericValue)) {
+    // If the expression is a rational, i.e. 1/2, 3/4, 5/6, etc.
+    // apply the rule to the numerator and the denominator
+    const [num, den] = expr.numericValue;
+    const numExpr = expr.engine.number(num);
+    const denExpr = expr.engine.number(den);
+    const numResult =
+      applyRule(rule, numExpr, substitution, options) ?? numExpr;
+    const denResult =
+      applyRule(rule, denExpr, substitution, options) ?? denExpr;
+    return numResult.div(denResult);
+  }
 
   let changed = false;
   if (expr.ops && options?.recursive) {
@@ -161,7 +598,7 @@ function applyRule(
 
   // If the condition doesn't match, the rule doesn't apply
   if (typeof condition === 'function' && !condition(sub, expr.engine))
-    return null;
+    return changed ? expr : null;
 
   // console.trace('apply rule ', id, 'to', expr.toString());
   // @debug
@@ -245,7 +682,24 @@ export function replace(
   return atLeastOneRule ? expr : null;
 }
 
-// @todo ['Alternatives', ...]:
-// @todo: ['Condition',...] : Conditional match
-// @todo: ['Repeated',...] : repeating match
-// @todo _x:Head or _x:RealNumbers
+/**
+ * For each rules in the rule set that match, return the `replace` of the rule
+ *
+ * @param rules
+ */
+export function matchRules(
+  expr: BoxedExpression,
+  rules: BoxedRuleSet,
+  sub: BoxedSubstitution
+): BoxedExpression[] {
+  const results: BoxedExpression[] = [];
+  for (const rule of rules) {
+    const r = applyRule(rule, expr, sub);
+    if (r === null) continue;
+    // Verify that the results are unique
+    if (results.some((x) => x.isSame(r))) continue;
+    results.push(r);
+  }
+
+  return results;
+}